:hammer: [layout] rename: types -> tags
